requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-docx
weasyprint
torch
whisperx @ git+https://github.com/m-bain/whisperX.git
